INS-482: move rolerecord SC
